Authorization added, code refactored, a lot of logic added to controllers

+++ Core/Interfaces/IUserActivityRepository.cs
@@ -5,6 +5,6 @@ namespace Core.Interfaces;
 public interface IUserActivityRepository
 {
     Task<UserActivity?> GetByIdAsync(int id);
-    Task<List<UserActivity>> GetByUserIdAsync(long userId);
+    Task<List<UserActivity>> GetByUserIdAsync(int userId);
     Task AddAsync(UserActivity activity);
 }
