+++ Core/Interfaces/IUserProgressRepository.cs
@@ -5,8 +5,8 @@ namespace Core.Interfaces;
 public interface IUserProgressRepository
 {
     Task<UserProgress?> GetByIdAsync(int id);
-    Task<UserProgress?> GetByUserAndContentAsync(long userId, int contentId);
-    Task<List<UserProgress>> GetByUserIdAsync(long userId);
+    Task<UserProgress?> GetByUserAndContentAsync(int userId, int contentId);
+    Task<List<UserProgress>> GetByUserIdAsync(int userId);
     Task AddAsync(UserProgress progress);
     Task UpdateAsync(UserProgress progress);
 }
